[refactor] #1 - Link, UseTerms 도메인을 policy -> external로 변경
