+++ src/main/java/com/napzak/domain/external/core/entity/UseTermsTableConstants.java
@@ -1,4 +1,4 @@
-package com.napzak.domain.policy.core.entity;
+package com.napzak.domain.external.core.entity;
 
 public class UseTermsTableConstants {
 	public static final String TABLE_USE_TERMS = "use_terms";
